+++ CHANGELOG.md
@@ -31,6 +31,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 * Invalid duration calculation for ``WAV``.
 * Properly handle invalid comment, involved people list, and TMCL frames
+* Properly support ID3v2.2 ``PIC`` frames.
 
 
 ## [0.9.0](https://github.com/thebigmunch/audio-metadata/releases/tag/0.9.0) (2020-03-25)